move bindings tests into file in tests/
